feat: allow to pass custom renderer
